Raise the light cap to 512 + add SUBFLOOR 988,793 by @SmokeMonsterPacks
200 was binding on everything, not just the one map that tripped it. The
procedural every-other-cell grid alone wants 225 fixtures on a 32x32, so
procgen levels have been silently losing their last 25 this whole time; a
community map arrived wanting 468. 512 covers both.

- MAX_LIGHTS 200 -> 512
- n_lights uint8_t -> uint16_t. 255 was a hard ceiling below the new cap, so
  the counter had to widen or 468 would have wrapped to 213
- the secondary's per-frame purge now covers only the LIVE entries rather
  than sizeof lights. The array is 4KB now, and purging all of it every frame
  would tax every map to serve the densest one

Adds the map that surfaced this. Its author had no way to see the limit: the
editor's Lights meter shipped one commit after the fly deploy, so the cap was
enforced while the meter that displays it was not. That's since deployed.

The map sits at 249 of 255 partition edges and chains to ALOHA COMPLEX via
the next: picker.

+++ sh_src/raycast.h
@@ -65,7 +65,7 @@ extern int g_lobby_ceiling;
  * from custom_maps.h, which includes this header — forward-declare instead. */
 struct cm_light_s;
 extern const struct cm_light_s *g_map_lights;
-extern uint8_t                  g_map_n_lights;
+extern uint16_t                 g_map_n_lights;
 /* When set, at least one low-ceiling crawlspace cell exists this map. */
 extern int g_lowceil_active;
 /* Per-cell ceiling height (255 = full open ceiling, lower = crawlspace slab).
